+++ test/Feature/Endpoint/Element/GetElementEndpoint/GetElementEndpointNormalNode200.test.ts
@@ -1,5 +1,7 @@
 import { expect } from 'chai';
-import { SinonSandbox, SinonStubbedInstance, assert, createSandbox } from 'sinon';
+import { HttpResponse, http } from 'msw';
+// eslint-disable-next-line import/no-unresolved
+import { setupServer } from 'msw/node';
 import { Container } from 'typedi';
 
 import GetElementEndpoint from '~/Endpoint/Element/GetElementEndpoint';
@@ -7,44 +9,53 @@ import { Logger } from '~/Service/Logger';
 import { WebSdkConfiguration } from '~/Service/WebSdkConfiguration';
 import { validateUuidFromString } from '~/Type/Definition/Uuid';
 
-import { mockServer } from '../../../../MockServer/mockServer';
+import { TestLogger } from '../../../TestLogger';
 
-describe('GetElementEndpoint normal node tests', () => {
-  let sandbox: SinonSandbox;
-  let mockedLogger: SinonStubbedInstance<Logger>;
-
-  beforeEach(() => {
-    sandbox = createSandbox();
-
-    mockServer.listen();
-
-    mockedLogger = sandbox.createStubInstance(Logger);
-    Container.set(Logger, mockedLogger);
-
-    Container.get(WebSdkConfiguration).setApiHost('http://mock-api');
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    Container.reset();
-    mockServer.close();
-  });
-
-  it('should load an accessible node', async () => {
-    const nodeUuid = validateUuidFromString('b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5');
-    const node = await Container.get(GetElementEndpoint).getElement(nodeUuid);
-
-    assert.calledOnceWithExactly(
-      mockedLogger.debug,
-      'Executing HTTP GET request against url http://mock-api/b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5 .',
+const mockServer = setupServer(
+  http.get('http://mock-api/b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5', () => {
+    return HttpResponse.json(
+      {
+        type: 'Data',
+        id: 'b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5',
+        data: {
+          created: '2023-10-06T20:27:56+00:00',
+          updated: '2023-10-06T20:27:56+00:00',
+          name: 'Test Data',
+        },
+      },
+      {
+        status: 200,
+        headers: {
+          'Content-Type': 'application/json; charset=utf-8',
+        },
+      },
     );
+  }),
+);
 
-    expect(node).to.have.keys('id', 'type', 'data');
-    expect(node).to.not.have.keys('start', 'end');
-    expect(node.id).to.equal(nodeUuid);
-    expect(node.type).to.equal('Data');
-    expect(node.data.created).to.be.instanceof(Date);
-    expect(node.data.updated).to.be.instanceof(Date);
-    expect(Object.keys(node.data)).to.have.lengthOf(3);
-  });
+const testLogger: TestLogger = new TestLogger();
+Container.set(Logger, testLogger);
+Container.get(WebSdkConfiguration).setApiHost('http://mock-api');
+
+test('GetElementEndpoint should handle normal node response', async () => {
+  mockServer.listen();
+  const uuid = validateUuidFromString('b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5');
+
+  const node = await Container.get(GetElementEndpoint).getElement(uuid);
+
+  expect(
+    testLogger.assertDebugHappened(
+      'Executing HTTP GET request against url http://mock-api/b1e85bf9-6a79-4e50-ae5a-ed49beac8cb5 .',
+    ),
+  ).to.be.true;
+
+  expect(node).to.have.keys('id', 'type', 'data');
+  expect(node).to.not.have.keys('start', 'end');
+  expect(node.id).to.equal(uuid);
+  expect(node.type).to.equal('Data');
+  expect(node.data.created).to.be.instanceof(Date);
+  expect(node.data.updated).to.be.instanceof(Date);
+  expect(Object.keys(node.data)).to.have.lengthOf(3);
+
+  mockServer.close();
 });